big update to python3 / maya 2022

+++ apps/tb_options.py
@@ -40,18 +40,18 @@ qtVersion = pm.about(qtVersion=True)
 if int(qtVersion.split('.')[0]) < 5:
     from PySide.QtGui import *
     from PySide.QtCore import *
-    from pysideuic import *
+    #from pysideuic import *
     from shiboken import wrapInstance
 else:
     from PySide2.QtWidgets import *
     from PySide2.QtGui import *
     from PySide2.QtCore import *
-    from pyside2uic import *
+    #from pyside2uic import *
     from shiboken2 import wrapInstance
 
 
 def getMainWindow():
-    return wrapInstance(long(omUI.MQtUtil.mainWindow()), QWidget)
+    return wrapInstance(int(omUI.MQtUtil.mainWindow()), QWidget)
 
 
 def showOptions():
